+++ plot/viewer/src/domain/Actor.ts
@@ -30,6 +30,7 @@ export class Actor implements BaseFields {
   readonly is_root!: boolean;
   readonly details_path!: string | null;
   readonly owner!: string | null;
+  readonly version!: string;
 
   readonly kind: "actor" = "actor";
 
@@ -82,6 +83,7 @@ export class Actor implements BaseFields {
       is_root: this.is_root,
       details_path: this.details_path,
       owner: this.owner,
+      version: this.version,
       kind: "actor",
       motivation: this.motivation,
       pain: this.pain,