+++ plot/viewer/src/domain/Category.ts
@@ -29,6 +29,7 @@ export class Category implements BaseFields {
   readonly is_root!: boolean;
   readonly details_path!: string | null;
   readonly owner!: string | null;
+  readonly version!: string;
 
   readonly kind: "category" = "category";
 
@@ -77,6 +78,7 @@ export class Category implements BaseFields {
       is_root: this.is_root,
       details_path: this.details_path,
       owner: this.owner,
+      version: this.version,
       kind: "category",
       theme: this.theme,
     };