+++ plot/viewer/src/domain/Identity.ts
@@ -32,6 +32,7 @@ export class Identity implements BaseFields {
   readonly is_root!: boolean;
   readonly details_path!: string | null;
   readonly owner!: string | null;
+  readonly version!: string;
 
   readonly kind: "identity" = "identity";
 
@@ -88,6 +89,7 @@ export class Identity implements BaseFields {
       is_root: this.is_root,
       details_path: this.details_path,
       owner: this.owner,
+      version: this.version,
       kind: "identity",
       description: this.description,
       do: this.do,