+++ plot/viewer/src/domain/Project.ts
@@ -32,6 +32,7 @@ export class Project implements BaseFields {
   readonly is_root!: boolean;
   readonly details_path!: string | null;
   readonly owner!: string | null;
+  readonly version!: string;
 
   readonly kind: "project" = "project";
 
@@ -67,6 +68,7 @@ export class Project implements BaseFields {
       is_root: this.is_root,
       details_path: this.details_path,
       owner: this.owner,
+      version: this.version,
       kind: "project",
     };
   }