+++ plot/viewer/src/domain/Service.ts
@@ -36,6 +36,7 @@ export class Service implements BaseFields {
   readonly is_root!: boolean;
   readonly details_path!: string | null;
   readonly owner!: string | null;
+  readonly version!: string;
 
   readonly kind: "service" = "service";
 
@@ -112,6 +113,7 @@ export class Service implements BaseFields {
       is_root: this.is_root,
       details_path: this.details_path,
       owner: this.owner,
+      version: this.version,
       kind: "service",
       target_side: this.target_side,
       what: this.what,